+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## Unreleased
 
+* `read-value` supports now `byte-array`.
+
 * Updated deps:
 
 ```clj
 [com.fasterxml.jackson.core/jackson-databind "2.9.8"] is available but we use "2.9.7"
 [com.fasterxml.jackson.datatype/jackson-datatype-jsr310 "2.9.8"] is available but we use "2.9.7"
-[com.fasterxml.jackson.datatype/jackson-datatype-joda "2.9.8"] is available but we use "2.9.7"
 ```
 
 ## 0.2.2 (2018-09-22)